Display chart of accounts in main window
This is temporary in terms of how the UX will work, but this will be
the first function I'll work on.

Currently this just displays a list of "default accounts", which is a
small list currently. The nominal codes have been thought about
briefly, but a proper convention should be determined in future when
the list is expanded.